Allow FATES-SPITFIRE to read lightning data from file
I replicated code used in CLM's BGC to read Fang Li's or FMEC's
lightning data for use in Fang Li's fire model.

I also turned on fates_seed_suppl in fates_params_default.cdl.
Initially I turned this on as a test because the model was crashing as a
result of the first way that I chose to pass the lightning data to
SPITFIRE. Turning on fates_seed_suppl didn't help, but I left it on
anyway.